requirements.txt
cfgv==3.4.0
distlib==0.3.9
filelock==3.18.0
identify==2.6.10
nodeenv==1.9.1
pika==1.3.2
platformdirs==4.3.8
pre_commit==4.2.0
python-decouple==3.8
PyYAML==6.0.2
virtualenv==20.31.2
